# ubkg-api-spec.yaml
openapi: 3.0.3
info:
  title: UBKG API
  description: The UBKG API provides REST endpoints that return results from queries against an instance of the [Unified Biomedical Knowledge Graph (UBKG)](https://ubkg.docs.xconsortia.org/). To prevent exceeding service timeouts or payload limits, some endpoints are throttled in both execution time and in the size of its response payload--e.g., to a maximum execution time of 15 seconds and a maximum payload of 9 MB.
  version: "2.0"
  contact:
    name: GitHub repository
    url: https://github.com/x-atlas-consortia/ubkg-api
servers:
  - url: https://ontology.api.hubmapconsortium.org/
    description: Production server
  - url: https://ontology-api.dev.hubmapconsortium.org/
    description: Development server
paths:
...
  /codes/{code_id}/concepts:
    get:
      operationId: codes_code_id_concepts_get
      summary: A Code node can have CODE relationships with (links to) multiple Concept nodes. For example, the Code with identifier SNOMEDCT_US:79962008 links to 5 different concepts. This endpoint returns the set of Concept nodes that have CODE relationships with the specified Code node.
      parameters:
        - name: code_id
          in: path
          required: true
          description: The identifier for the Code node, in format _SAB_:_CODE_ (or SAB%3Acode), where _SAB_ is an identifier for the source of the code. (%3A = colon)
          schema:
            type: string
            example: SNOMEDCT_US:79962008
      responses:
        '200':
          description: An array of Concept nodes that have CODE relationships with the specified Code node
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptDetail'
        '404':
          description: no Concept nodes for the specified Code
        '5XX':
          description: Unknown error
  /concepts/{concept_id}/codes:
    get:
      operationId: concepts_concept_id_codes_get
      summary: A Concept node can have CODE relationships with (links to) multiple Code nodes. This endpoint returns information on the set of Code nodes that link directly to the specified Concept, subject to constraints specified in parameters.
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The identifier for the Concept (also known as the Concept Unique Identifier, or CUI)
          schema:
            type: string
            example: C0678222
        - name: sab
          in: query
          required: false
          description: The set of sources (SABs) to which to limit the set of Codes that link to the specified Concept. Can be either a list of values delimited with URL-encoded commas (e.g., ?sab=SAB1%2CSAB2%2SAB3) or with individual key-value pairs (e.g., ?sab=SAB1&sab=SAB2&SAB=SAB3).
          schema:
            type: array
            uniqueItems: true
            items:
              type: string
              example: MEDLINEPLUS,MSH,MEDCIN
            default: []
      responses:
        '200':
          description: An array of code_ids in format _SAB_:_CODE for the Codes that link to the specified Concept, subject to constraints specified in parameters.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: CHV 0000052645
        '400':
          description: invalid parameter name
        '404':
          description: no Codes for the specified parameters (_concept_id_ and _sab_)
        '5XX':
          description: Unknown error
  /concepts/{concept_id}/concepts:
    get:
      operationId: concepts_concept_id_concepts_get
      summary: Returns the set of Concepts that have direct relationships with the specified Concept.
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The concept identifier (also known as Concept Unique Identifier, or CUI)
          schema:
            type: string
            example: C0006142
      responses:
        '200':
          description: An array of information on Concepts that have relationships with the specified Concept
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SabRelationshipConceptTerm'
        '404':
          description: no Concepts that have relationships with the specified Concept
        '5XX':
          description: Unknown error
  /concepts/{concept_id}/definitions:
    get:
      operationId: concepts_concept_id_definitions_get
      summary: Returns a list of Definitions for a Concept. A Definition is usually defined for a Code by a source (SAB); however, in the UBKG, the Definition is associated with the Code's linked Concept. Thus, a Concept may have multiple Definitions.
      parameters:
        - name: concept_id
          description: Identifier (also known as a Concept Unique Identifier) for a Concept.
          in: path
          required: true
          schema:
            type: string
            example: C0006142
      responses:
        '200':
          description: An array of Definitions associated with the Concept.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SabDefinition'
        '404':
          description: No Definitions associated with the Concept _query_concept_id_.
        '5XX':
          description: Unknown error
  /concepts/{identifier}/nodes:
    get:
      operationId: concepts_search_nodes_get
      summary: Returns a set of "Concept node" objects for Concept nodes that "match" an identifier. A Concept node in the UBKG is the origin of a subgraph that links the Concept node to a set of Code, Term, Definition, and Semantic Type nodes. A Concept node object represents information extracted from all of the nodes in the Concept node's subgraph. A "match" is the union of the set of EXACT matches on text-based properties of nodes in the subgraph that originates from a Concept node.
      parameters:
        - name: identifier
          in: path
          required: true
          description: The string to match against properties related to a Concepts. Related properties include the preferred term for the Concept; the CUI for the Concept; the CodeID for a Code linked to the Concept; and terms for a Code linked to a Concept. Special characters in search should be URL-encoded (%20 = space; %3A = colon, %2C = comma).
          schema:
            type: string
            example: Cells
      responses:
        '200':
          description: The set of Concept node objects that match the search term.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptNodes'
        '404':
          description: No Concepts that match the search term
        '5XX':
          description: Unknown error
  /concepts/{concept_id}/paths/expand:
    get:
      operationId: concepts_paths_expand_get
      summary: Returns a set of paths originating from the specified Concept, subject to constraints specified in parameters.
      parameters:
        - name: concept_id
          in: path
          required: true
          description: The identifier (also known as Concept Unique Identifier, or CUI) for the Concept from which paths originate.
          schema:
            type: string
            example: C2720507
        - name: sab
          in: query
          required: true
          description: The set of sources (SABs) to which to limit the relationships in the set of expanded paths. Can be either a list of values delimited with URL-encoded commas (e.g., ?sab=SAB1%2CSAB2%2SAB3) or with individual key-value pairs (e.g., ?sab=SAB1&sab=SAB2&SAB=SAB3).
          schema:
            type: string
            example: SNOMEDCT_US
        - name: rel
          in: query
          required: true
          description: The set of relationship types to which to limit the relationships in the set of expanded paths. Can be either a list of values delimited with URL-encoded commas (e.g., ?rel=relationship_type11%2Crelationship_type2) or individual key-value pairs (e.g., ?rel=relationship_type1&rel=relationship_type2)
          schema:
            type: string
            example: isa
        - name: mindepth
          in: query
          required: false
          description: The minimum number of hops in the returned set of expanded paths--e.g., mindepth=2 directs the endpoint to return paths that have at least 2 hops. The minimum must be non-negative and less than the value of _maxdepth_. The default value is 1. Large values of mindepth may result in long queries that exceed the execution time limits.
          schema:
            type: string
            example: 1
        - name: maxdepth
          in: query
          required: false
          description: The maximum number of hops in the returned set of expanded paths--e.g., maxdepth=4 directs the endpoint to return paths that have no more than 4 hops. The maximum must be non-negative. The default value is _mindepth_ + 2. Large values of maxdepth may result in long queries that exceed the execution time limits.
          schema:
            type: string
            example: 3
        - name: skip
          in: query
          required: false
          description: the number of paths to skip in the returned set. The value must be non-negative. The default value is 0.
          schema:
            type: string
            example: 0
        - name: limit
          in: query
          required: false
          description: the maximum number of paths to return. The value must be non-negative. The default value is the maximum number of rows specified by configuration.
          schema:
            type: string
            example: 10
      responses:
        '200':
          description: Return the Concepts in all paths originating from the specified Concept, subject to constraints. Each path is an ordered set of objects representing hops away from the originating concept. Each hop in a path represents a relationship between two concepts. For example, the path (A)-[r]->(B)-[r]->(C) has 2 hops between Concept C and Concept A.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptPaths'
        '400':
          description: invalid parameter name; missing required parameter name; non-numeric parameter value; negative value; mindepth > maxdepth
        '404':
          description: No Concepts with paths originating from the specified concept and set of parameters. Includes responses from queries with execution times that exceed the maximum.
        '5XX':
          description: Unknown error
...
  /terms/{term_id}/concepts:
    get:
      operationId: terms_term_id_concepts_get
      summary: Returns a list of Concept nodes with preferred terms that exactly match the specified string.
      parameters:
        - name: term_id
          in: path
          required: true
          description: The string to match
          schema:
            type: string
            example: Breast cancer
      responses:
        '200':
          description: An array of Concepts with preferred terms that exactly match the specified string
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: C0794811
        '404':
          description: No Concepts with terms matching the specified string.
        '5XX':
          description: Unknown error
components:
  schemas:
...
    ConceptDetail:    # Schema name
      type: object
      description: Information on a Concept node.
      properties:
        concept:
          type: string
          description: Concept Unique Identifier (CUI) for the Concept node
          example: C0006142
        prefterm:
          type: string
          description: Preferred term for the Concept node.
          example: Malignant neoplasm of breast
...
    SabDefinition: # Schema name
      type: object
      description: A Definition associated with a Concept.
      properties:
        sab:
          type: string
          description: Source of a Definition.
          example: NCI
        definition:
          type: string
          description: The definition text.
          example: A primary or metastatic malignant neoplasm involving the breast. The vast majority of cases are carcinomas arising from the breast parenchyma or the nipple. Malignant breast neoplasms occur more frequently in females than in males.
    SabRelationshipConceptTerm: # Schema name
      type: object
      description: Information on a Concept that has a relationship with another Concept
      properties:
        sab:
          type: string
          description: The source of the relationship, such as a vocabulary or ontology, that asserts the relationship between concepts. Although relationships are usually described in terms of Codes, they actually exist between the Concepts that are linked to the Codes. For example, if an ontology SAB1 shows that code A has relationship r with code B, the associated concepts for A and B will have a relationship with type=r and SAB=SAB1.
          example: ICD10AM
        relationship:
          type: string
          description: the relationship type--e.g., "isa", "property_of"
          example: CHD
        concept:
          type: string
          description: the identifier for the Concept, also known as the Concept Unique Identifier (CUI)
          example: C0006826
        prefterm:
          type: string
          description: The preferred term for the Concept. Note that the preferred term for a Concept may be different than the preferred term for a Code that links to the Concept.
          example: Malignant Neoplasms
...